fix(hpb): Deprecate multiple HPBs and fake-clustering
