+++ lib/AppInfo/Application.php
@@ -109,7 +109,7 @@ use OCA\Talk\Search\UnifiedSearchFilterPlugin;
 use OCA\Talk\Settings\Personal;
 use OCA\Talk\SetupCheck\BackgroundBlurLoading;
 use OCA\Talk\SetupCheck\FederationLockCache;
-use OCA\Talk\SetupCheck\RecommendCache;
+use OCA\Talk\SetupCheck\HighPerformanceBackend;
 use OCA\Talk\SetupCheck\RecordingBackend;
 use OCA\Talk\SetupCheck\SIPConfiguration;
 use OCA\Talk\Share\Listener as ShareListener;
@@ -338,7 +338,7 @@ class Application extends App implements IBootstrap {
 
 		$context->registerTeamResourceProvider(TalkTeamResourceProvider::class);
 
-		$context->registerSetupCheck(RecommendCache::class);
+		$context->registerSetupCheck(HighPerformanceBackend::class);
 		$context->registerSetupCheck(FederationLockCache::class);
 		$context->registerSetupCheck(RecordingBackend::class);
 		$context->registerSetupCheck(SIPConfiguration::class);
